Big changes on Schema Migration Table. New name: sharpmigrations. ... Versions are logged one by one and not just the last version number. Since versions are logged one by one, we can now run old migrations if they weren't run (common case when merging migrations created by different devs on source control)

+++ Sharp.Migrations/Runners/ConsoleRunner.cs
@@ -76,7 +76,7 @@ namespace Sharp.Migrations.Runners {
 			_targetVersion = GetIntFromConsole(string.Format(ASK_FOR_VERSION, runner.LastVersionNumber));
 		}
 
-        protected void ShowCurrentVersion(int version) {
+        protected void ShowCurrentVersion(long version) {
             Console.WriteLine(SHOW_CURRENT_VERSION, version);
         }
